Fix clang warnings

+++ src/hexen/p_mobj.c
@@ -537,7 +537,7 @@ void P_XYMovement(mobj_t * mo)
                     {
                         case MT_CENTAUR:
                         case MT_CENTAURLEADER:
-                            if (abs(angle - BlockingMobj->angle) >> 24 > 45)
+                            if (abs((int) angle - (int) BlockingMobj->angle) >> 24 > 45)
                                 goto explode;
                             if (mo->type == MT_HOLY_FX)
                                 goto explode;
